load and save files transactionally
